Do not send messages to a subscriber until it is ready.
When a subscriptions is created it is attached to a channel before it starts replaying messages.  There is a small window where STAN can forward incoming messages to a subscriber before resending these messages in the replay.  If these messages are then resent in the replay, they will be duplicates without the resend flag set, and initially sent out of order.  This fix adds a started flag to prevent new messages from being sent until a subscriber is ready.

Various existing tests were failing at times in Travis.  Equivalent java tests were failing on a more consistent basis.