Fixed: sub2 should not receive more than half what is sent
There is no guarantee that sub2 receives only 1 or 2 message.
The publisher sends, server acks the publisher and pick a queue
sub to send the message to. If the client's callback is delayed,
the server may send several messages to both sub1 and sub2 since
they would both accumulate the same number of acks.