Remove SubscriptionCloseRequest protocol
This is related to PR #115 and server side [PR](https://github.com/nats-io/nats-streaming-server/pull/203)
Noticed that we could have used UnsubscribeRequest since content
of that one and SubscriptionCloseRequest were same.
Also, the server was actually implemented with UnsubscribeRequest
and works fine. What matters is the subject the protocol is sent to.