[FIXED] Possible duplicate processing of a PubAck
It is possible that the client timeout at the same time that it
receives the pubAck from the server. It is acceptable to have
the client report a failure because it timed-out, even though
the ack has just been received, but it is not correct to invoke
the AckHandler twice for the same guid.

Also, the issue would cause serious problems for synchronous
Publish() calls. See issue report for details.

Resolves #130