De-couple Msg from private Subscriber implementation
Currently, Msg's Ack() method assumes that its Subcriber is typed
as the internal subscription structure and will panic is such is
not the case. This makes using a mock subscriber unsafe and
impossible without type checking code forks.

Since the code within Msg's Ack() is intrinsically tied to the
subscriber struct, it made sense to convert that code into a
method of subscriber. Msg will now be acknowledged based on the
definition of its Subscriber.