[ADDED] Ability to set a queue group for stan-bench
You can now specify `-qgroup` to specify a queue group.
Note that since the number of messages expected is for the whole
group, then the first member to make the total number of
received messages reach the goal, it will emit a new message to
make all other members break out of their loop.

Resolves #200

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>