Skip code that would try to send CloseRequest on failed connect
We were calling c.Close() inside Connect() for any failure
condition. The problem is that Close() is trying to send a
close request, which does not make sense (and will fail with
invalid subject internally since closeRequests subject is "").

So for any failure before processing the ConnectResponse will
now call a function that does proper cleanup without trying to
send the close request protocol.
After getting a ConnectionResponse OK we *must* call Close()
since we need to let the server know that we are actually
closing (no connection is returned to the user, so user can't
Close() explicitly).

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>