[IMPROVED] Added wrong cluster ID hint in connect timeout error
When a client calls stan.Connect() with a cluster ID that is
different that the server's cluster ID, the request will timeout.
This is expected since the server does not receive the connect
request at all, so the client times out waiting for the reply.

We are adding a hint in the error message which may help some users:
`stan: connect request timeout (possibly wrong cluster ID?)`

Resolves #285

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>