Cleanup some code
- Move all ping related fields into own pingInfo struct.
- Use connection closed boolean to check/set closed state instead
  of setting nc to nil. That makes sc.nc immutable and not nil
  and so code using that doesn't have to capture/check.
- Add a pubAckCloseChan that is closed during connection close
  to prevent publish call from being possibly blocked.

Regarding last point, we have tests and I tried to add a new one
but it would not catch the issue unless I would add a sleep in
publishAsync() after check of connection close and before sending
to pubAckChan.
The race is that if the connection is fully closed between these
two lines then if there are enough go routines trying to send to
the pubAckChan or MaxPubAcksInflight is low, the channel could be
full which would block the publishAsync() call.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>