[FIXED] Fail of publish ack on connection close
We can't invoke the ack handler from the connection close code
since there would be a possible deadlock if user code is calling
any connection method.

This is based out of #297 but without the change on how pub acks
are timed-out (still use timers). So this PR has only tests fixes
and the fix for failing pub acks on connection close.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>