[ADDED] NatsOptions to configure the NATS connection on stan.Connect()
This allows users that don't want to create their own NATS connection
to configured as desired, to pass the NATS options they would use
to create the NATS connection.
With that, the streaming connection will create and own the NATS
connection but will be configured based on the provided NATS options,
except for some options that are overridden by the library.

Resolve #354

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>