[ADDED] AllowCloseRetry option
In PR #360, we changed the default behavior of connection Close()
in that if the library failed to sent the close protocol or received
its response, the Close() could be retried. In order to do that,
the underlying NATS connection would not be closed.

This has side-effects in that the NATS connection would be left
opened or possibly reconnect when an user would have previously
called Close(), not check error, and expect NATS connection to
be closed.

The introduction of this option allow the default behavior to
remain the same, and users that do want to be able to call Close()
again on a failed call will need to set this option when creating
the connection.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>